Handle NaN in permutation tests
Pervious patch assigned NaN values to test statistics for invalid permutations.
This patch allows the permutation testing routine to skip these invalid permutations
and adjust the effective number of permutations. With this, the output will report
the actual permutations used as well as valid summary statistics for permutation
testing procedure such as the median of statistics.