Implement batched output behavior for eqtlbma_bf
Break analysis into batches so that results will be
written to disk upon completion of each batch, and all
data of the batch will be deleted after results are written.
Previously the program will not output any results until all
runs are complete. That requires a lot memory when there are
many gene-snp pairs to analyze. This patch is more memory effecient.
Note that it is inevitable that the progress bar behavior have to
be changed because unlike previous implementation where the association
test part and permutation part use separate progress bars, the new
implementation combines the two progress bars into one.